ml/requirements.txt
scikit-learn
supabase
pandas
numpy
matplotlib
seaborn
joblib
python-dotenv
